+++ Polygon_mesh_processing/include/CGAL/Polygon_mesh_processing/remesh.h
@@ -21,7 +21,7 @@
 #ifndef CGAL_POLYGON_MESH_PROCESSING_REMESH_H
 #define CGAL_POLYGON_MESH_PROCESSING_REMESH_H
 
-#include <CGAL/Polygon_mesh_processing/internal/remesh_impl.h>
+#include <CGAL/Polygon_mesh_processing/internal/Isotropic_remeshing/remesh_impl.h>
 
 #include <CGAL/Polygon_mesh_processing/internal/named_function_params.h>
 #include <CGAL/Polygon_mesh_processing/internal/named_params_helper.h>
